inspector/requirements.txt
fastapi
uvicorn
openai
azure-identity
azure-keyvault-secrets